HAI-719 provide retrieved projects from context to all children elements within Map of HankeMap;
Refactor API logic away from HankeLayer, move that data to a provider
